+++ .env.example
@@ -1,0 +1,5 @@
+# DIRECTORIES
+DATASET_LOCATION=../../0-botnetCollection/datasets
+CTU_DIR=/CTU-13-Dataset
+NCC_DIR=/NCC
+URL_EXTRACTED_DATA=../extract/
