+++ main.py
@@ -1,0 +1,7 @@
+"""Writen By: M. Aidiel Rachman Putra"""
+"""Organization: Net-Centic Computing Laboratory | Institut Teknologi Sepuluh Nopember"""
+
+import utilities.menuManagement as menu
+
+if __name__ == "__main__":
+    menu.mainMenu()
