+++ utilities/dataLoader.py
@@ -1,5 +1,5 @@
 import pandas as pd
-import pickle
+from utilities.globalConfig import DATASET_LOCATION, CTU_DIR, NCC_DIR
 
 #all CTU-13 dataset scenarios
 ctuOnline = {
@@ -18,8 +18,8 @@ ctuOnline = {
   'scenario13': 'https://mcfp.felk.cvut.cz/publicDatasets/CTU-Malware-Capture-Botnet-54/detailed-bidirectional-flow-labels/capture20110815-3.binetflow',
 }
 
-datasetLocation = '../../../0-botnetCollection/datasets'
-ctuLoc = '/CTU-13-Dataset'
+datasetLocation = DATASET_LOCATION
+ctuLoc = CTU_DIR
 ctu = {
   'scenario1': datasetLocation+ctuLoc+'/1/capture20110810.binetflow',
   'scenario2': datasetLocation+ctuLoc+'/2/capture20110811.binetflow',
@@ -36,7 +36,7 @@ ctu = {
   'scenario13': datasetLocation+ctuLoc+'/13/capture20110815-3.binetflow',
 }
 
-nccLoc = '/NCC'
+nccLoc = NCC_DIR
 ncc = {
   'scenario1': datasetLocation+nccLoc+'/scenario_dataset_1/dataset_result.binetflow',
   'scenario2': datasetLocation+nccLoc+'/scenario_dataset_2/dataset_result.binetflow',
