+++ utilities/menuManagement.py
@@ -1,9 +1,6 @@
-"""Writen By: M. Aidiel Rachman Putra"""
-"""Organization: Net-Centic Computing Laboratory | Institut Teknologi Sepuluh Nopember"""
-
-from dataLoader import *
-from extract import *
-from simulate import *
+from utilities.dataLoader import *
+from src.extract import *
+from src.simulate import *
 import time
 import os
 
@@ -14,41 +11,7 @@ listMenu = [
     "EXIT"
 ]
 
-#extract
-def extract():
-    print("\n******************")
-    print("Dataset Option: ")
-    print("1. CTU-13")
-    print("2. NCC")
-    print("******************")
-    datasetChoose = input("Select Dataset:")
-    if(datasetChoose == "1"):
-        datasetName = ctu
-        stringDatasetName = 'ctu'
-    else:
-        datasetName = ncc
-        stringDatasetName = 'ncc'
-    for i in range(1,14):
-        selectedScenario = 'scenario'+str(i)
-        df = loadDataset(datasetName, selectedScenario)
-        splitActivity(stringDatasetName, df, i)
-
-#simulation
-def simulate():
-    simulation(1, 3,10,1,2,7,12,5)
-    analytics(1)
-    simulation(2, 10,11,1,2,9,6,13)
-    analytics(2)
-    simulation(3, 3,4,9,8,12,5,13)
-    analytics(3)
-
-#merging sensors
-def merge():
-    mergingAllSensors()
-    # analytics('s-all')
-
-#menu
-def menu():
+def banner():
     print("\n\n================================================================")
     print("======| Botnet Simultaneous Dataset Generator |=================")
     print("================================================================\n")
@@ -56,7 +19,8 @@ def menu():
     for i, menu in enumerate(listMenu):
         print(str(i+1)+". "+menu)
     print("\n================================================================")
-    choose = input("Enter Menu: ")
+
+def executor(choose):
     print("============| Processing Menu: "+choose+". "+listMenu[int(choose)-1]+" |========================\n")
     if(choose == "1"):
         extract()
@@ -84,6 +48,8 @@ def menu():
         time.sleep(3) # adding 3 seconds time delay
         os.system("clear")
 
-if __name__ == "__main__":
-   while(True):
-    menu()
+#menu
+def mainMenu():
+    banner()
+    choose = input("Enter Menu: ")
+    executor(choose)
